Refactor pyard module  - moved ARD into ard.py  - moved other functions/constants to misc.py  - fix for lru_cache, pyard.init()  - add test to check cache changeability
